Apply cosmetic changes and fix build warnings
- Run clang-format on packages/services/Car/cpp/evs.
- Explicitly casting types a few arguments for StringAppendF() call in
  HalCamera::toString().
- Correct the order of member variables' initializations.

Fix: 345535095
Fix: 345514992
Test: Treehugger
Flag: EXEMPT refactor

+++ cpp/evs/manager/aidl/src/HalCamera.cpp
@@ -378,8 +378,7 @@ ScopedAStatus HalCamera::deliverFrame(const std::vector<BufferDesc>& buffers) {
 
         // Adding skipped capture requests back to the queue.
         std::lock_guard<std::mutex> lock(mFrameMutex);
-        mNextRequests.insert(mNextRequests.end(),
-                             std::make_move_iterator(puntedRequests.begin()),
+        mNextRequests.insert(mNextRequests.end(), std::make_move_iterator(puntedRequests.begin()),
                              std::make_move_iterator(puntedRequests.end()));
         mFrameOpInProgress = false;
         mFrameOpDone.notify_all();
@@ -402,8 +401,7 @@ ScopedAStatus HalCamera::deliverFrame(const std::vector<BufferDesc>& buffers) {
         }
 
         // Adding skipped capture requests back to the queue.
-        mNextRequests.insert(mNextRequests.end(),
-                             std::make_move_iterator(puntedRequests.begin()),
+        mNextRequests.insert(mNextRequests.end(), std::make_move_iterator(puntedRequests.begin()),
                              std::make_move_iterator(puntedRequests.end()));
         mFrameOpInProgress = false;
         mFrameOpDone.notify_all();
@@ -577,8 +575,9 @@ std::string HalCamera::toString(Stream configuration, const char* indent) {
                   "%srotation: 0x%X\n\n",
                   indent, double_indent.data(), configuration.id, double_indent.data(),
                   configuration.width, double_indent.data(), configuration.height,
-                  double_indent.data(), configuration.format, double_indent.data(),
-                  configuration.usage, double_indent.data(), configuration.rotation);
+                  double_indent.data(), static_cast<unsigned int>(configuration.format),
+                  double_indent.data(), static_cast<unsigned long>(configuration.usage),
+                  double_indent.data(), static_cast<unsigned int>(configuration.rotation));
 
     return streamInfo;
 }
